feat: implement comprehensive k6 performance testing infrastructure with InfluxDB and Grafana integration
- Add AI model search performance testing suite with multiple test scenarios (smoke, load, stress, spike)
- Implement JWT-based authentication in k6 tests with proper cookie handling
- Configure InfluxDB v1.8 integration for k6 metrics storage with native k6 output support
- Create comprehensive Grafana dashboard for k6 performance monitoring with 9 visualization panels
- Add custom k6 metrics: search_error_rate, search_response_time, active_search_users, search_request_count
- Implement URLSearchParams compatibility fix for k6 runtime environment
- Configure monitoring stack with Prometheus, InfluxDB, and Grafana datasource provisioning
- Add test execution scripts with health checks and multiple output formats
- Fix API response structure validation from 'data' to 'response' field format
- Update Docker Compose monitoring configuration for InfluxDB v1.8 compatibility

+++ src/main/java/com/example/nomodel/subscription/domain/model/Subscription.java
@@ -35,7 +35,7 @@ public class Subscription {
     public Long getId() { return id; }
     public PlanType getPlanType() { return planType; }
     public String getDescription() { return description; }
-    public Integer getPeriod() { return period; }
+    public Long getPeriod() { return period; }
     public Integer getDailyLimit() { return dailyLimit; }
     public BigDecimal getPrice() { return price; }
     public Integer getSelfMadeModelNum() { return selfMadeModelNum; }
